--- imagersite/imager_images/templates/albums_and_photos/images.html
@@ -1,11 +1,0 @@
-{% extends 'generic/base.html' %}
-{% load thumbnail %}
-
-{% block  %}
-
-{% if photos %}
-{% for photo in photos %}
-<a href="photo.image.url"><img class="library" src="{{photo.image.url}}" alt=""></a>
-{% endfor %}
-
-{% endblock  %}
