+++ imagersite/imager_images/views.py
@@ -1,5 +1,6 @@
-from django.shortcuts import render, redirect
+from django.shortcuts import render, redirect, get_object_or_404
 from .models import Photo, Album
+from imager_profile.models import ImagerProfile
 
 
 def photos_view(request):
@@ -17,6 +18,58 @@ def albums_view(request):
         'albums': album
     }
     return render(request, 'imager_images/album.html', albums)
+
+
+def library_view(request, username=None):
+    owner = False
+
+    if not username:
+        username = request.user.get_username()
+        owner = True
+        if username == '':
+            return redirect('home')
+
+    profile = get_object_or_404(ImagerProfile, user__username=username)
+    album = Album.objects.filter(user__username=username)
+    photos = Photo.objects.filter(album__user__username=username)
+
+    if not owner:
+        photos = Photo.objects.filter(published='PUBLIC')
+        album = Album.objects.filter(published='PUBLIC')
+
+    context = {
+        'profile': profile,
+        'album': album,
+        'photos': photos
+    }
+
+    return render(request, 'imager_images/library.html', context)
+
+
+def photo_view(request, photo_id=None):
+    owner = False
+    import pdb; pdb.set_trace()
+    if not username:
+        username = request.user.get_username()
+        owner = True
+        if username == '':
+            return redirect('home')
+
+    profile = get_object_or_404(ImagerProfile, user__username=username)
+    album = Album.objects.filter(user__username=username)
+    photos = Photo.objects.filter(album__user__username=username)
+
+    if not owner:
+        photos = Photo.objects.filter(published='PUBLIC')
+        album = Album.objects.filter(published='PUBLIC')
+
+    context = {
+        'profile': profile,
+        'album': album,
+        'photos': photos
+    }
+
+    return render(request, 'imager_images/library.html', context)
