+++ imagersite/imager_profile/views.py
@@ -27,34 +27,3 @@ def profile_view(request, username=None):
     }
 
     return render(request, 'imager_profile/user.html', context)
-
-
-def library_view(request, username=None):
-    owner = False
-
-    if not username:
-        username = request.user.get_username()
-        owner = True
-        if username == '':
-            return redirect('home')
-
-    profile = get_object_or_404(ImagerProfile, user__username=username)
-    album = Album.objects.filter(user__username=username)
-    photos = Photo.objects.filter(album__user__username=username)
-
-    if not owner:
-        photos = Photo.objects.filter(published='PUBLIC')
-        album = Album.objects.filter(published='PUBLIC')
-
-    context = {
-        'profile': profile,
-        'album': album,
-        'photos': photos
-    }
-
-    return render(request, 'imager_profile/library.html', context)
-
-
-
-
-
