+++ imagersite/imagersite/templates/generic/base.html
@@ -22,6 +22,7 @@
         {% if user.is_authenticated %}
         <li><a href="{% url "profile" %}">Profile</a></li>
         <li><a href="{% url "photos" %}">Pictures</a></li>
+        <li><a href="{% url "library" %}">Library</a></li>
         <li><a href="{% url "albums" %}">Albums</a></li>
         <li><a href="{% url "auth_logout" %}">Logout</a></li>
         {% else %}
